Added martintest to main Lost-in-Time project. Re-did some folders. A folder for main menu resources, one for character animation.

+++ Project-Enjoy/Lost-in-Time/Player/player.cpp
@@ -1,11 +1,7 @@
 #include <SFML/Graphics.hpp>
 #include "player.h"
 
-Player::Player(float x, float y, const Config& config, sf::RenderWindow& window)
-        : x(x), y(y),
-          config(config),
-          window(window)
-
+Player::Player(int x, int y, const Config& config, sf::RenderWindow& window) : x(x), y(y), config(config), window(window)
 {
     character.setSize(sf::Vector2f(size, size));
     character.setFillColor(sf::Color::Red);
